+++ src/modules/bzip2/common.h
@@ -15,11 +15,11 @@ namespace archive{ namespace impl {
         using stream = bz_stream;
         static void fill_out(stream& z,void* base,size_t len) {
             z.next_out = static_cast<char*>(base);
-            z.avail_out = len;
+            z.avail_out = static_cast<unsigned int>(len);
         }
         static void fill_in(stream& z,void* base,size_t len) {
             z.next_in = reinterpret_cast<char*>(base);
-            z.avail_in = len;
+            z.avail_in = static_cast<unsigned int>(len);
         }
         static bool has_in(stream& z) {
             return z.avail_in;
